Rename src/app/api.py -> src/app/main.py
